Move iovecs test logic to testlib
